Move type definitions out of GraphQL tests into own package
